Forward has been completed.

+++ utility/params.hpp
@@ -3,11 +3,11 @@
 
 namespace FlashAttention::utility {
     struct Params {
-        int   batch_size;
-        int   num_heads;
-        int   seq_len;
-        int   head_dim;
-        float scale = 0;
+        int64_t batch_size;
+        int64_t num_heads;
+        int64_t seq_len;
+        int64_t head_dim;
+        float   scale = 0;
 
         CUTE_HOST_DEVICE signed long one_head_size() const noexcept {
             return seq_len * head_dim;
